api/requirements.txt
fastapi
uvicorn
pydantic
google-auth
requests
psycopg
python-multipart
pillow
pytesseract
opencv-python-headless
numpy
